requirements.txt
aiogram==2.11.2
mysql-connector-python==2.2.9
python-dateutil==2.8.2